+++ public-info/src/main/java/com/bank/publicinfo/entity/Certificate.java
@@ -1,12 +1,9 @@
 package com.bank.publicinfo.entity;
 
-import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-
 import javax.persistence.*;
-import java.math.BigInteger;
 
 @Entity
 @Getter
@@ -18,7 +15,7 @@ public class Certificate {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private BigInteger id;
+    private Long id;
 
     @Column(name = "photo", nullable = false)
     private byte[] photo;
