+++ public-info/src/main/java/com/bank/publicinfo/repository/AtmRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigInteger;
 
-public interface AtmRepository extends JpaRepository<Atm, BigInteger> {
+public interface AtmRepository extends JpaRepository<Atm, Long> {
 }
